1.2

+++ app/Http/Controllers/UserController.php
@@ -3,31 +3,33 @@
 namespace App\Http\Controllers;
 
 use App\User;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use phpDocumentor\Reflection\Types\Integer;
-use function MongoDB\BSON\toJSON;
+use ReflectionClass;
+
 
 class UserController extends Controller
 {
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
      */
     public function index() {
         Log::info('Show all users');
         $user = User::all();
-        return response()->json($user);
+        return response()->json($user,200);
     }
 
     /**
      * Show the form for creating a new resource.
      *
      * @param Request $request
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
      */
     public function create(Request $request) {
        //  User::create($request->all);
@@ -43,7 +45,7 @@ class UserController extends Controller
         $user->api_token= $user->generateToken();
 
         $user->save();
-        return response()->json($user);
+        return response()->json($user,201);
     }
 
     /**
@@ -61,12 +63,15 @@ class UserController extends Controller
      * Display the specified resource.
      *
      * @param $id
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
      */
     public function show($id){
-        Log::info('Find user by id '.$id);
-        $user = User::find($id);
-        return response()->json($user);
+        if (!(User::find($id)))return response()->json(array('message'=>'Not Found','code'=>404),404);
+        else {
+            Log::info('Find user by id ' . $id);
+            $user = User::find($id);
+            return response()->json($user,200);
+        }
     }
 
     /**
@@ -74,16 +79,19 @@ class UserController extends Controller
      *
      * @param Request $request
      * @param $id
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
      */
     public function edit(Request $request, $id) {
-        Log::info('Update user'.$id);
-        $user = User::findOrFail($id);
-        $user->update($request->all());
-        return response()
-          // ->setStatusCode(201, 'The resource is created successfully!')
-           //->header('Content-Type','application/json')
-          ->json($user);
+        if (!(User::find($id)))return response()->json(array('message'=>'Not Found','code'=>404),404);
+        else {
+            Log::info('Update user' . $id);
+            $user = User::findOrFail($id);
+            $user->update($request->all());
+            return response()
+                // ->setStatusCode(201, 'The resource is created successfully!')
+                //->header('Content-Type','application/json')
+                ->json($user,201);
+        }
     }
 
     /**
@@ -101,11 +109,14 @@ class UserController extends Controller
      * Remove the specified resource from storage.
      *
      * @param $id
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
      */
     public function destroy($id){
-        Log::info('Delete user'.$id);
-        $user= User::find($id)->delete();
-        return response()->json($user);
+        if (!(User::find($id)))return response()->json(array('message'=>'Not Found','code'=>404),404);//return response()->json(array(Res::HTTP_NOT_FOUND,Res::$statusTextsEn[404]));
+        else {
+            Log::info('Delete user'.$id);
+            $user = User::find($id)->delete();
+            return response()->json(array('message'=>'No Content','code'=>204),204);
+        }
     }
 }
